updated with suspense wrapper for vercel fix
