+++ frontend-ui/src/app/workouts/start/page.jsx
@@ -1,16 +1,12 @@
-"use client"
-//initial page when user clicks start (not the during page)
-//will have workout id passed
-import {  useSearchParams } from 'next/navigation';
+import React, { Suspense } from 'react';
+import StartWorkoutPage from '../../../components/workouts/StartWorkoutPage';
 
-const StartWorkoutPage = () => {
-    const searchParams = useSearchParams();
-    const id = searchParams.get('id'); // Access the query parameter
-    return (
-      <div>
-        <p>Start workout {id}</p>
-      </div>
-    )
-  };
-  
-  export default StartWorkoutPage;
+const StartWorkoutPageWrapper = () => {
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <StartWorkoutPage />
+    </Suspense>
+  );
+};
+
+export default StartWorkoutPageWrapper;
